Refactor invoke tasks
Favor importing and running other tasks over running
dependent tasks via shell.

Add docstrings to tasks that were missing them.

Move build and install to 'package' submodule.

Update documentation.
