+++ tasks/test.py
@@ -1,11 +1,11 @@
 import subprocess
 from pathlib import Path
+from time import sleep
 
 from invoke import task
 
-from tasks.housekeeping import clean
-
 PROJECT_DIR = Path(__file__).parent.parent
+DOCKER_TAG = "test-anybadge:latest"
 
 
 @task
@@ -19,9 +19,8 @@ def local(c):
 
 
 def build_test_docker_image():
-    subprocess.run(
-        f"(cd docker/test && docker build . -t test-anybadge:latest)", shell=True
-    )
+    print("Building test docker image... ")
+    subprocess.run(f"(cd docker/test && docker build . -t {DOCKER_TAG})", shell=True)
 
 
 @task
@@ -29,11 +28,14 @@ def docker(c):
     """Run dockerised tests."""
     print("Running containerised tests...")
 
-    subprocess.run("invoke clean", shell=True)
-    subprocess.run("invoke build", shell=True)
+    from tasks.housekeeping import clean
+    from tasks.package import build
+
+    clean(c)
+    build(c)
     build_test_docker_image()
     subprocess.run(
-        f"docker run -v {PROJECT_DIR}:/app test-anybadge:latest /work/run_docker_tests.sh",
+        f"docker run --rm -v {PROJECT_DIR}:/app {DOCKER_TAG} /work/run_docker_tests.sh",
         shell=True,
     )
 
@@ -43,13 +45,17 @@ def pypi(c, version="latest"):
     """Run tests against Pypi version."""
     print("Running tests against pypi version...")
 
+    from tasks.housekeeping import clean
+
     clean(c)
+
     test_files = PROJECT_DIR / Path("test_files")
     test_files.mkdir(exist_ok=True)
 
     build_test_docker_image()
+    print("Running tests in docker image... ")
     subprocess.run(
-        f"docker run -e VERSION={version} -v {test_files.absolute()}:/test_files test-anybadge:latest /work/run_pypi_tests.sh",
+        f"docker run --rm -e VERSION={version} -v {test_files.absolute()}:/test_files {DOCKER_TAG} /work/run_pypi_tests.sh",
         shell=True,
     )
 
@@ -58,6 +64,8 @@ def pypi(c, version="latest"):
 def cli(c, version="latest"):
     """Run CLI tests against currently installed version."""
     print("Running tests against currently installed version...")
+
+    from tasks.housekeeping import clean
 
     clean(c)
     test_files = PROJECT_DIR / Path("test_files")
